refactor: Move token logic to TokensClient
